fix(deploy/agent): always automountServiceAccountToken if not explicitly disable
Since #109 the token of the pod is required for the API server checks.
Previously the tokens were only automounted when the "k8s exporter"
feature is enabled.
Now the tokens are always mounted and the ClusterRole,
ClusterRoleBinding & ServiceAccount objects are always created. The
ClusterRole only has rules if the "k8s exporter" feature is enabled.